requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
beautifulsoup4==4.12.2
blobfile==2.1.1
bpy==4.0.0
Brotli
certifi
charset-normalizer
clip==0.2.0
comm==0.2.2
Cython==3.0.11
decorator==5.1.1
diffusers==0.29.2
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.13
filelock
flash-attn==2.6.3
fsspec==2024.6.1
ftfy==6.2.3
gmpy2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna
importlib_metadata==8.4.0
ipython==8.26.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2
jiter==0.5.0
jupyterlab_widgets==3.0.13
lxml==4.9.4
MarkupSafe
matplotlib-inline==0.1.7
mkl-fft
mkl-random
mkl-service==2.4.0
mpmath
networkx
numpy
open_clip_torch==2.26.1
openai==1.43.0
openai-clip==1.0.1
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PySocks
PyYAML
rarfile==4.2
regex==2024.7.24
requests
safetensors==0.4.4
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy
tiktoken==0.5.2
timm==1.0.9
tokenizers==0.15.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.38.0
triton==3.0.0
typing_extensions
urllib3
wcwidth==0.2.13
widgetsnbextension==4.0.13
xformers==0.0.27.post2+cu118
zipp==3.20.1
zstandard==0.23.0